requirements.txt
phidata
openai
groq
yfinance
duckduckgo-search
python-dotenv